Move timestamp and duration helpers into a dedicated common timestamp module
