+++ src/windows/common/precomp.h
@@ -146,6 +146,7 @@ Abstract:
 #include "conncheckshared.h"
 #include "helpers.hpp"
 #include "string.hpp"
+#include "timestamp.hpp"
 #include "filesystem.hpp"
 #include "Localization.h"
 #include "wslutil.h"
